Corrected Windows compatibility errors.

+++ include/glaze/network/core.hpp
@@ -46,7 +46,7 @@ namespace glz::net
 {
 #ifdef _WIN32
    using event_handle_t = HANDLE;
-   using ssize_t = int64_t;
+   using ssize_t = int32_t;
 #else
    using event_handle_t = int;
 #endif
